+++ R/Drivers.R
@@ -47,7 +47,7 @@ jdbcDrivers <- new.env()
 #' - Spark (Databricks): V3.3.3
 #' - Snowflake: V4.2.0
 #' - BigQuery: v0.10.0
-#' - InterSystems IRIS: v3.10.2
+#' - InterSystems IRIS: v3.11.0
 #' 
 #' @return Invisibly returns the destination if the download was successful.
 #' @export
@@ -97,7 +97,7 @@ downloadJdbcDrivers <- function(dbms, pathToDriver = Sys.getenv("DATABASECONNECT
     5,spark,DatabricksJDBC-3.3.3.zip,https://databricks-bi-artifacts.s3.us-east-2.amazonaws.com/simbaspark-drivers/jdbc/3.3.3/
     6,snowflake,snowflake-jdbc-4.2.0.jar,https://repo1.maven.org/maven2/net/snowflake/snowflake-jdbc/4.2.0/
     7,bigquery,google-cloud-bigquery-jdbc-0.10.0-all.jar,https://repo1.maven.org/maven2/com/google/cloud/google-cloud-bigquery-jdbc/0.10.0/
-    8,iris,intersystems-jdbc-3.10.2.jar,https://repo1.maven.org/maven2/com/intersystems/intersystems-jdbc/3.10.2/"
+    8,iris,intersystems-jdbc-3.11.0.jar,https://repo1.maven.org/maven2/com/intersystems/intersystems-jdbc/3.11.0/"
   )
 
   if (dbms == "all") {